requirements.txt
django
requests
redis
Pillow
tenacity
git+https://github.com/zxh326/django-simple-serializer.git
gevent
django-simditor
sentry_sdk
mysqlclient